+++ test/transfers_page_batch_actions_test.dart
@@ -12,6 +12,7 @@ import 'package:remote_storage/models/paged_listings.dart';
 import 'package:remote_storage/models/remote_storage_config.dart';
 import 'package:remote_storage/models/s3_objects.dart';
 import 'package:remote_storage/models/share_record.dart';
+import 'package:remote_storage/models/system_proxy_info.dart';
 import 'package:remote_storage/models/trash_item.dart';
 import 'package:remote_storage/models/transfer_job.dart';
 import 'package:remote_storage/models/sync_profile.dart';
@@ -105,6 +106,9 @@ class _TransfersPageFakeApi implements RemoteStorageGateway {
 
   @override
   Future<BridgeBuildInfo> getBuildInfo() async => const BridgeBuildInfo();
+
+  @override
+  Future<SystemProxyInfo?> resolveSystemProxy() async => null;
 
   @override
   Future<String> installApp({