+++ test/widget_test.dart
@@ -15,6 +15,7 @@ import 'package:remote_storage/models/remote_storage_config.dart';
 import 'package:remote_storage/models/paged_listings.dart';
 import 'package:remote_storage/models/s3_objects.dart';
 import 'package:remote_storage/models/share_record.dart';
+import 'package:remote_storage/models/system_proxy_info.dart';
 import 'package:remote_storage/models/trash_item.dart';
 import 'package:remote_storage/models/transfer_job.dart';
 import 'package:remote_storage/models/sync_profile.dart';
@@ -164,7 +165,10 @@ class _FakeApi implements RemoteStorageGateway {
       const AuthSessionState.desktop();
 
   @override
-  Future<BridgeBuildInfo> getBuildInfo() async => const BridgeBuildInfo();
+ Future<BridgeBuildInfo> getBuildInfo() async => const BridgeBuildInfo();
+
+  @override
+  Future<SystemProxyInfo?> resolveSystemProxy() async => null;
 
   @override
   Future<String> installApp({